app/requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
ecdsa==0.19.1
fastapi==0.115.12
fastapi-filter==2.0.1
fastapi-pagination==0.13.1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.4.8
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.40
starlette==0.46.2
trio==0.30.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
